DEPLOY-136
- Moved code for URL-based message to home component
- Removed custom home component
