+++ examples/hello-world/ui/src/app/app.module.ts
@@ -6,13 +6,10 @@ import { HomeComponent } from './home/home.component';
 import { AppComponent } from './app.component';
 
 import {HttpModule} from '@angular/http';
-import { CustomHomeComponent } from './customHome/customHome.component';
-
 
 const appRoutes: Routes = [
-  { path: '', component: HomeComponent },
-  { path: 'hello/welcome', redirectTo: 'hello/welcome', pathMatch: 'full' },
-  { path: 'hello/:id', component: CustomHomeComponent }
+  { path: '', redirectTo: 'hello/welcome', pathMatch: 'full' },
+  { path: 'hello/:id', component: HomeComponent }
 ];
 
 @NgModule({
@@ -27,8 +24,7 @@ const appRoutes: Routes = [
   ],
   declarations: [
     AppComponent,
-    HomeComponent,
-    CustomHomeComponent
+    HomeComponent
   ],
   providers: [],
   bootstrap: [AppComponent]
