--- examples/hello-world/ui/src/app/customHome/customHome.component.html
@@ -1,1 +1,0 @@
-<h3>{{msg}}</h3>
