+++ examples/hello-world/ui/src/app/home/home.component.ts
@@ -10,20 +10,34 @@ import { ActivatedRoute } from '@angular/router';
 })
 export class HomeComponent {
 
+  private id;
+  private sub: any;
   private apiUrl;
   data: any ={};
   msg;
-  constructor(private http:Http,
-              appConfig: AppConfigService) {
-    this.apiUrl = appConfig.get('backEndHost') + "/hello/welcome";
-    this.getResponse();
+
+  constructor(private route: ActivatedRoute, private http:Http,
+        appConfig: AppConfigService) {
+    this.apiUrl = appConfig.get('backEndHost') + '/';
   }
 
-  getResponse() {
-    return this.http.get(this.apiUrl).
-      map((res: Response) => res.json()).subscribe(data => {
-        this.msg=data.value;
-        this.data = data
-      })
+  private ngOnInit() {
+     this.sub = this.route.params.subscribe(params => {
+       this.id = params['id'];
+      });
+     this.getResponse(this.id);
   }
+
+  getResponse(id) {
+     this.apiUrl += id;
+     return this.http.get(this.apiUrl).
+       map((res: Response) => res.json()).subscribe(data => {
+         this.msg=data.value;
+         this.data = data;
+         console.log(data.status)
+       },
+       err => {
+         this.msg = 'ERROR: Something went wrong!';
+       })
+   }
 }
