+++ examples/hello-world/ui/src/assets/app.config.json
@@ -1,3 +1,3 @@
 {
-    "backEndHost": "http://HELLO_BACKEND_HOST:HELLO_BACKEND_PORT/"
+    "backEndHost": "http://HELLO_BACKEND_HOST:HELLO_BACKEND_PORT/hello"
 }
